Refactor twin_networks into subfiles
