+++ diffrend/torch/GAN/discriminator_networks/__init__.py
@@ -1,0 +1,5 @@
+from diffrend.torch.GAN.discriminator_networks._netD_64 import _netD_64
+from diffrend.torch.GAN.discriminator_networks._netD_256 import _netD_256
+from diffrend.torch.GAN.discriminator_networks._netD import _netD
+from diffrend.torch.GAN.discriminator_networks._netD2 import _netD2
+from diffrend.torch.GAN.discriminator_networks.DCGAN_D import DCGAN_D
