+++ diffrend/torch/GAN/generator_networks/__init__.py
@@ -1,0 +1,5 @@
+from diffrend.torch.GAN.generator_networks._netG_mlp import _netG_mlp
+from diffrend.torch.GAN.generator_networks._netG import _netG
+from diffrend.torch.GAN.generator_networks.DCGAN_G import DCGAN_G
+from diffrend.torch.GAN.generator_networks.DCGAN_G2 import DCGAN_G2
+from diffrend.torch.GAN.generator_networks._netG_resnet import _netG_resnet
